// types/globalTypes.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

//==============================================================
// START Enums and Input Objects
//==============================================================

export enum CountEnum {
  CHAT_URL = "CHAT_URL",
  PHONE_NUMBER = "PHONE_NUMBER",
  SMS_NUMBER = "SMS_NUMBER",
  URL = "URL",
  VIEW = "VIEW",
}

export enum LocalityEnum {
  COUNTY = "COUNTY",
  LOCATION = "LOCATION",
  PROVINCE = "PROVINCE",
  REGION = "REGION",
  STATE = "STATE",
}

/**
 * Autogenerated input type of CountryIncrementCountMutation
 */
export interface CountryIncrementCountMutationInput {
  code: string;
  clientMutationId?: string | null;
}

/**
 * Autogenerated input type of CountrySubdivisionIncrementCountMutation
 */
export interface CountrySubdivisionIncrementCountMutationInput {
  code: string;
  countryCode: string;
  clientMutationId?: string | null;
}

/**
 * Autogenerated input type of OrganizationIncrementCountMutation
 */
export interface OrganizationIncrementCountMutationInput {
  slug: string;
  count: CountEnum;
  clientMutationId?: string | null;
}

//==============================================================
// END Enums and Input Objects
//==============================================================